Move k6 js module to internal as preparations for v1.0.0
This wasn't previously done due to the exported `ErrCheckInInitContext`
being used in a couple of extensions that should likely come up with
their own errors.

Related to #4133
