+++ internal/js/modules/k6/k6_test.go
