+++ internal/js/runner_test.go
@@ -35,13 +35,13 @@ import (
 	"go.k6.io/k6/errext"
 	"go.k6.io/k6/internal/execution"
 	"go.k6.io/k6/internal/execution/local"
+	"go.k6.io/k6/internal/js/modules/k6"
 	k6metrics "go.k6.io/k6/internal/js/modules/k6/metrics"
 	"go.k6.io/k6/internal/js/modules/k6/ws"
 	"go.k6.io/k6/internal/lib/testutils"
 	"go.k6.io/k6/internal/lib/testutils/httpmultibin"
 	"go.k6.io/k6/internal/lib/testutils/httpmultibin/grpc_testing"
 	"go.k6.io/k6/internal/lib/testutils/mockoutput"
-	"go.k6.io/k6/js/modules/k6"
 	k6http "go.k6.io/k6/js/modules/k6/http"
 	"go.k6.io/k6/lib"
 	_ "go.k6.io/k6/lib/executor" // TODO: figure out something better
